refactor: rename ai-spacing/in-extension-pages.ts to model.ts
Move the page-side model download (LanguageModel.create) in from
options.ts so model.ts owns every page-side LanguageModel call:
availability probe, can-run check, and download start. The old name
described which half of the classifier split by context the file
was; model.ts names what it actually holds, the model's lifecycle as
the popup and options pages see it. PAGE_MODEL_LANGUAGES no longer
needs to be exported now that its only other caller lives in the
same file.
