+++ browser-extensions/chrome/src/ai-spacing/model.ts
@@ -2,7 +2,7 @@
 // warning is logged for extension pages only, never for the service worker (measured 2026-09-02), so declaring a language on the pages' calls is what clears it. Nothing here is misattested:
 // no page call ever produces model output, one probes availability and the other exists only to start the browser-wide download. The classifier, which is the session that actually prompts, stays
 // undeclared on purpose -- see in-service-worker.ts. `en` is a member of the supported set (en/ja/es/de/fr); no zh variant is, and declaring one makes availability() report unavailable.
-export const PAGE_MODEL_LANGUAGES: LanguageModelExpected[] = [{ type: 'text', languages: ['en'] }];
+const PAGE_MODEL_LANGUAGES: LanguageModelExpected[] = [{ type: 'text', languages: ['en'] }];
 
 // `unsupported` is ours, not an API value: the browser has no Prompt API at all, so availability() cannot even be asked
 export type AiModelAvailability = Availability | 'unsupported';
@@ -20,4 +20,10 @@ export async function getAiModelAvailability(): Promise<AiModelAvailability> {
 // False when the model can never run here (no API, or this machine fails its requirements), so the toggle is shown off and disabled. A model that is merely not downloaded yet still counts
 export function canAiModelRun(availability: AiModelAvailability) {
   return availability !== 'unsupported' && availability !== 'unavailable';
+}
+
+// The download is browser-wide and outlives the page that starts it, so the session only exists to start it
+export async function startAiModelDownload() {
+  const session = await LanguageModel.create({ expectedOutputs: PAGE_MODEL_LANGUAGES });
+  session.destroy();
 }
